Update provision command and host version type

+++ contracts/core/Host.sol
@@ -1,13 +1,13 @@
 // SPDX-License-Identifier: GPL-3.0-only
 pragma solidity ^0.8.33;
 
-import {AccessControl} from "./Access.sol";
-import {Authorize} from "../commands/admin/Authorize.sol";
-import {Unauthorize} from "../commands/admin/Unauthorize.sol";
-import {Relocate} from "../commands/admin/Relocate.sol";
-import {HostAnnouncedEvent} from "../events/HostAnnounced.sol";
-import {IHostDiscovery} from "../interfaces/IHostDiscovery.sol";
-import {ensureHost} from "../utils/Ids.sol";
+import { AccessControl } from "./Access.sol";
+import { Authorize } from "../commands/admin/Authorize.sol";
+import { Unauthorize } from "../commands/admin/Unauthorize.sol";
+import { Relocate } from "../commands/admin/Relocate.sol";
+import { HostAnnouncedEvent } from "../events/HostAnnounced.sol";
+import { IHostDiscovery } from "../interfaces/IHostDiscovery.sol";
+import { ensureHost } from "../utils/Ids.sol";
 
 abstract contract HostDiscovery is HostAnnouncedEvent, IHostDiscovery {
     function announceHost(uint id, uint blocknum, uint16 version, string calldata namespace) external {
@@ -16,7 +16,7 @@ abstract contract HostDiscovery is HostAnnouncedEvent, IHostDiscovery {
 }
 
 abstract contract Host is Authorize, Unauthorize, Relocate {
-    constructor(address cmdr, uint8 version, string memory namespace) AccessControl(cmdr) {
+    constructor(address cmdr, uint16 version, string memory namespace) AccessControl(cmdr) {
         if (cmdr == address(0) || cmdr == address(this) || cmdr.code.length == 0) return;
         IHostDiscovery(cmdr).announceHost(host, block.number, version, namespace);
     }
